requirements.txt
astroid==1.6.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2-binary==2.7.4
pylint==1.8.2
six==1.11.0
wrapt==1.10.11
